Address review: harden fence close and improve coverage
Respond to code-review findings on the code-block pipe-line fix:

- Restrict the closing-fence blank check to ASCII spaces and tabs, so a
  no-break space (U+00A0) or form feed (U+000C) after a marker is literal
  content rather than a valid close, matching CommonMark.
- Remove the dead duplicate regression from the un-wired
  `tests/table/` suite; the authoritative copy already lives in the
  compiled `tests/table_continuations.rs` target.
- Split the fence-tracker transition-logging tests into
  `src/wrap/tests/fence_tracker_logging.rs` so both modules stay within
  the 400-line file limit.
- Add property coverage for the fence info-string invariant and for lone
  single-cell table passthrough.
- Log the `parse_and_validate` bail-outs with stable `reason`,
  `row_count`, and `has_separator` fields for production diagnosis.
- Document the code-block and fence-closing behaviour in the users' guide.
