+++ src/wrap/tests.rs
@@ -6,6 +6,7 @@
 mod blockquote;
 mod classify_block;
 mod fence_tracker;
+mod fence_tracker_logging;
 mod inline_wrapping;
 mod link_ref_regex;
 mod link_reference_definitions;
